More mir-opt test fixes for big-endian platforms
A number of mir-opt test cases are failing again on big-endian
platforms, due to endian-specific binary encodings in the output
files that are being checked by the test.  To fix those, one of
the following two strategies is used:

a) Where a binary encoded value directly originates from some
immediate constant in the test case source code, change that
constant to an endian-invariant value (like 0x11000011).

b) Where that is not easily possible or would change the essence
of what the test is checking for, disable the test on big-endian
platforms.

Fixes (again) https://github.com/rust-lang/rust/issues/105383.